# docs/swagger.yaml
basePath: /
definitions:
...
  updater.Status:
    properties:
      asset_name:
        type: string
      build_time:
        type: string
      can_upgrade:
        type: boolean
      current_version:
        type: string
      latest_version:
        type: string
      platform:
        type: string
      prerelease:
        type: boolean
      published_at:
        type: string
      release_name:
        type: string
      release_notes:
        type: string
      release_url:
        type: string
      update_available:
        type: boolean
      upstream_repository:
        type: string
    type: object
...
  util.ResponseAny:
    properties:
      data: {}
      error_msg:
        example: ""
        type: string
    type: object
info:
  contact:
    name: Wavelet
    url: https://github.com/Rain-kl/Wavelet
  description: Wavelet 平台后端 API，提供用户认证、系统配置、任务调度等通用功能。
  license:
    name: AGPL-3.0
    url: https://www.gnu.org/licenses/agpl-3.0.html
  title: Wavelet API
  version: 1.0.0
paths:
...
  /api/v1/admin/update:
    get:
      description: 从系统配置指定的 GitHub 上游仓库查询最新兼容 Release，并与当前服务版本比较
      produces:
      - application/json
      responses:
        "200":
          description: 更新状态
          schema:
            allOf:
            - $ref: '#/definitions/util.ResponseAny'
            - properties:
                data:
                  $ref: '#/definitions/updater.Status'
              type: object
        "401":
          description: 未登录
          schema:
            $ref: '#/definitions/util.ResponseAny'
        "403":
          description: 无管理员权限
          schema:
            $ref: '#/definitions/util.ResponseAny'
        "500":
          description: 查询失败
          schema:
            $ref: '#/definitions/util.ResponseAny'
      security:
      - SessionCookie: []
      summary: 获取应用更新状态
      tags:
      - admin
  /api/v1/admin/update/apply:
    post:
      description: 下载当前平台对应的 GitHub Actions Release 资产，替换当前二进制并重启进程
      produces:
      - application/json
      responses:
        "200":
          description: 升级已准备并即将重启
          schema:
            $ref: '#/definitions/util.ResponseAny'
        "400":
          description: 当前版本不可升级
          schema:
            $ref: '#/definitions/util.ResponseAny'
        "401":
          description: 未登录
          schema:
            $ref: '#/definitions/util.ResponseAny'
        "403":
          description: 无管理员权限
          schema:
            $ref: '#/definitions/util.ResponseAny'
        "500":
          description: 升级准备失败
          schema:
            $ref: '#/definitions/util.ResponseAny'
      security:
      - SessionCookie: []
      summary: 下载并应用应用更新
      tags:
      - admin
...
securityDefinitions:
  SessionCookie:
    in: cookie
    name: session
    type: apiKey
swagger: "2.0"